OP-268/PipX Bootloader - Changed the size of BL def. New naming convention for the linker files: link_(board_name)_(density)_(bl usage).ld
git-svn-id: svn://svn.openpilot.org/OpenPilot/trunk@2401 ebee16cc-31ac-478f-84a7-5cbb03baadba

+++ flight/Bootloaders/PipXtreme/inc/op_dfu.h
@@ -37,7 +37,7 @@ typedef struct
 	uint32_t 	sizeOfCode;
 	uint8_t 	sizeOfDescription;
 	uint8_t 	BL_Version;
-	uint8_t 	devID;
+	uint16_t 	devID;
 	DeviceType  devType;
 	uint32_t 	FW_Crc;
 }Device;
